create add plugin function
