+++ src/packages/auth/package.json
@@ -26,6 +26,7 @@
 		"@apollo/server": "4.10.4",
 		"@exogee/graphweaver": "workspace:*",
 		"@exogee/graphweaver-scalars": "workspace:*",
+		"@exogee/graphweaver-server": "workspace:*",
 		"@exogee/logger": "workspace:*",
 		"@simplewebauthn/server": "10.0.0",
 		"apollo-server-errors": "3.3.1",
