+++ src/packages/auth/src/authentication/methods/magic-link.ts
@@ -11,6 +11,7 @@ import { requireEnvironmentVariable } from '../../helper-functions';
 import { BackendProvider, ResolverOptions, graphweaverMetadata } from '@exogee/graphweaver';
 import { AuthenticationType } from '../../types';
 import { AuthenticationBaseEntity } from '../entities';
+import { BaseAuthMethod } from './base-auth-method';
 
 const config = {
 	rate: {
@@ -43,12 +44,14 @@ export interface MagicLinkOptions {
 // For now this is just a uuid
 const createToken = randomUUID;
 
-export class MagicLink {
+export class MagicLink extends BaseAuthMethod {
 	private provider: MagicLinkProvider;
 	private getUser: (username: string) => Promise<UserProfile<unknown>>;
 	private sendMagicLink: (url: URL, magicLink: MagicLinkEntity) => Promise<boolean>;
 
 	constructor({ provider, getUser, sendMagicLink }: MagicLinkOptions) {
+		super();
+
 		this.getUser = getUser;
 		this.sendMagicLink = sendMagicLink;
 		this.provider = provider;
